refactor: move otel-test into k8s folder
