+++ tests/k8s/images/test-instrumentation/Dockerfile
