+++ tests/k8s/otel/otel-dev/otlphttp-mirror.yml
