+++ tests/k8s/otel/otel-test/otel-collector.yml
