+++ tests/k8s/otel/test_k8s_instrumentation.py
@@ -6,13 +6,13 @@ from typing import List, Dict
 from jina.logging.logger import JinaLogger
 
 from jina import Flow
-from tests.k8s_otel.kind_wrapper import KindClusterWrapperV2
-from tests.k8s_otel.util import parse_string_jaeger_tags, get_last_health_check_data
+from tests.k8s.kind_wrapper import KindClusterWrapper
+from tests.k8s.otel.util import parse_string_jaeger_tags, get_last_health_check_data
 
 
 @pytest.mark.asyncio
 @pytest.mark.timeout(1800)
-async def test_flow_resource_labeling(tmpdir, otel_test_namespace: str, k8s_cluster_v2: KindClusterWrapperV2):
+async def test_flow_resource_labeling(tmpdir, otel_test_namespace: str, k8s_cluster: KindClusterWrapper):
     NAMESPACE = 'test-flow-resource-labeling'
     dump_path = os.path.join(tmpdir, NAMESPACE)
     logger = JinaLogger(NAMESPACE)
@@ -35,10 +35,10 @@ async def test_flow_resource_labeling(tmpdir, otel_test_namespace: str, k8s_clus
     flow.to_kubernetes_yaml(dump_path, k8s_namespace=NAMESPACE)
 
     # Deploy flow
-    k8s_cluster_v2.deploy_from_dir(dir=dump_path, namespace=NAMESPACE)
+    k8s_cluster.deploy_from_dir(dir=dump_path, namespace=NAMESPACE)
 
     # Make client requests
-    with k8s_cluster_v2.port_forward('svc/gateway', NAMESPACE, svc_port=8080) as gateway_port:
+    with k8s_cluster.port_forward('svc/gateway', NAMESPACE, svc_port=8080) as gateway_port:
         from jina import Client
         [docs async for docs in Client(port=gateway_port, asyncio=True).post("/")]
 
@@ -46,7 +46,7 @@ async def test_flow_resource_labeling(tmpdir, otel_test_namespace: str, k8s_clus
     await asyncio.sleep(60)
 
     # Check Jaeger API
-    with k8s_cluster_v2.port_forward('svc/jaeger', otel_test_namespace, svc_port=16686) as jaeger_port:
+    with k8s_cluster.port_forward('svc/jaeger', otel_test_namespace, svc_port=16686) as jaeger_port:
         try:
             # Gateway
             trace_data = get_last_health_check_data(jaeger_port=jaeger_port, service_name='gateway')
@@ -67,7 +67,7 @@ async def test_flow_resource_labeling(tmpdir, otel_test_namespace: str, k8s_clus
             logger.error(trace_data)
             raise e
 
-    with k8s_cluster_v2.port_forward('svc/prometheus', otel_test_namespace, svc_port=9090) as prometheus_port:
+    with k8s_cluster.port_forward('svc/prometheus', otel_test_namespace, svc_port=9090) as prometheus_port:
         try:
             # Check Prometheus Labels
             prometheus_labels: List[str] = requests.get(f'http://localhost:{prometheus_port}/api/v1/labels').json()['data']
