+++ tests/k8s/otel/util.py
